+++ components/profile/ProfileDetail.vue
@@ -150,7 +150,7 @@ import TabItem from '../shared/TabItem.vue'
 import Identity from '../identity/IdentityIndex.vue'
 import ItemsGrid from '../items/ItemsGrid/ItemsGrid.vue'
 import ProfileGrid from './ProfileGrid.vue'
-import ProfileActivity from './ProfileActivity.vue'
+import ProfileActivity from './ProfileActivitySummery.vue'
 import FilterButton from './FilterButton.vue'
 import ChainDropdown from '../common/ChainDropdown.vue'
 import OrderByDropdown from './OrderByDropdown.vue'
